All comments addressed.
