+++ WhaleSpotting/react-app/src/apiModels/ApiEnums.ts
@@ -1,0 +1,25 @@
+export enum Species {
+    AtlanticWhiteSidedDolphin = 1,
+    CaliforniaSeaLion,
+    DallsPorpoise,
+    GrayWhale,
+    HarborPorpoise,
+    HarborSeal,
+    Humpback,
+    Minke,
+    NorthernElephantSeal,
+    Orca,
+    Other,
+    PacificWhiteSidedDolphin,
+    SeaOtter,
+    SouthernElephantSeal,
+    StellerSeaLion,
+    Unknown,
+}
+
+export enum OrcaType {
+    NorthernResident = 1,
+    Offshore,
+    SouthernResident,
+    Transient,
+}
